+++ src/main.rs
@@ -1,5 +1,5 @@
 mod threaded_archiver;
-mod listing;
+mod directory_listing;
 mod web;
